third_party/WebKit/LayoutTests/inspector/elements/styles-2/keyframes-rules.html

<html>
<head>
<style>
@keyframes fadeout {
    from { background-color: black; }
    to { background-color: white; }
}

* {
    background-color: papayawhip;
}
</style>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script>
function test()
{
    InspectorTest.runTestSuite([
        function testInit(next)
        {
            InspectorTest.selectNodeAndWaitForStyles("inspected", next);
        },

        function testDumpStyles(next)
        {
            InspectorTest.dumpSelectedElementStyles(true);
            next();
        }
    ]);
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that source data is extracted correctly from stylesheets with @keyframes rules.
</p>

<div id="inspected" style="background-color: white;">Content</div>
</body>
</html>
